feat: add engine abstraction
