+++ public/engines/demo.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <meta charset="utf-8">
+  <title>Engine Demo</title>
+  <style>html,body{margin:0;height:100%}canvas{width:100%;height:100%}</style>
+</head>
+<body>
+  <canvas id="demoCanvas"></canvas>
+  <script type="module" src="./demo.js"></script>
+</body>
+</html>
