+++ public/engines/demo.js
@@ -1,0 +1,5 @@
+import { initGlobe, render } from '../globe.js';
+
+const canvas = document.getElementById('demoCanvas');
+initGlobe(canvas);
+render();
